+++ src/paredit/commands.ts
@@ -147,3 +147,14 @@ export function rewrapSquare(doc: EditableDocument, isMulti: boolean) {
 export function rewrapParens(doc: EditableDocument, isMulti: boolean) {
   return paredit.rewrapSexpr(doc, '(', ')', isMulti ? doc.selections : [doc.selections[0]]);
 }
+
+export async function squeeze(
+  doc: EditableDocument,
+  isMulti: boolean,
+  onRange: (doc: EditableDocument, range: [number, number]) => Promise<void> = async () => {
+    // Default empty implementation
+  }
+) {
+  // TODO: support multi-cursor
+  return paredit.squeezeSexpr(doc, onRange, doc.selections[0].active);
+}